pyproject.toml
[project]
name = "atlassian-local-cli"
version = "0.4.0"
description = "CLI to interact with Confluence and Jira from the command line"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atlassian-python-api>=3.41.0",
    "html2text>=2025.4.15",
    "markdown>=3.7",
    "python-dotenv>=1.1.0",
]

[project.scripts]
atlassian-local-cli = "atlassian_local_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atlassian_local_cli"]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
    "pytest>=8.0",
    "pytest-cov>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/atlassian_local_cli"]

[tool.coverage.report]
exclude_lines = [
    "if __name__",
    "pragma: no cover",
]
